Rename the devx-api types file
