+++ src/app/services/reducers/devxApi-reducers.ts
@@ -1,5 +1,5 @@
 import { IAction } from '../../../types/action';
-import { IDevxAPI } from '../../../types/IDevxAPI';
+import { IDevxAPI } from '../../../types/devx-api';
 import { DEVX_API_URL } from '../graph-constants';
 import { SET_DEVX_API_URL_SUCCESS } from '../redux-constants';
